requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
requests
pydantic
httpx
python-dotenv
selenium==4.18.1
beautifulsoup4==4.12.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.22.0  # Optional but speeds up fuzzywuzzy

# annotated-types==0.7.0
# anyio==4.8.0
# black==25.1.0
# certifi==2025.1.31
# cffi==1.17.1
# charset-normalizer==3.4.1
# click==8.1.8
# cryptography==44.0.2
# fastapi==0.115.11
# h11==0.14.0
# idna==3.10
# iniconfig==2.0.0
# isort==6.0.1
# mypy==1.15.0
# mypy-extensions==1.0.0
# packaging==24.2
# pathspec==0.12.1
# pdfminer.six==20231228
# pdfplumber==0.11.5
# Pillow==11.1.0
# platformdirs==4.3.6
# pluggy==1.5.0
# pycparser==2.22
# pydantic==2.10.6
# pydantic-core==2.27.2
# pypdfium2==4.30.1
# pytest==8.3.5
# python-multipart==0.0.20
# sniffio==1.3.1
# starlette==0.46.0
# typing-extensions==4.12.2
# urllib3==2.3.0
# uvicorn==0.34.0

# # Development dependencies
# pytest
# black
# isort
# mypy